moved user settings retrieval to  settings.py, all error handling will be kept to this file
